Moved the original methods to _meteorMethods
I tried to keep the constraint definition.meteorMethods ==
schema.meteorMethods, but had to break it for the wrapped meteorMethods,
as they get initiated after the definition is parsed

Also added _meteorMethods as reservedKeyword
