+++ lib/core/reserved_keywords.js
@@ -1,5 +1,6 @@
 const reservedNames = [
   '_isNew',
+  '_meteorMethods',
   'copy',
   'dispatchEvent',
   'get',
