+++ lib/modules/meteorMethods/hooks/init_definition.js
@@ -1,5 +1,5 @@
 function onInitDefinition(definition, className) {
-	definition.meteorMethods = {};
+	definition._meteorMethods = {};
 };
 
 export default onInitDefinition;
