+++ lib/modules/meteorMethods/hooks/merge_definitions.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
 function onMergeDefinitions(targetDefinition, sourceDefinition, ClassName) {
-	_.each(sourceDefinition.meteorMethods, function(method, methodName) {
-		targetDefinition.meteorMethods[methodName] = method;
+	_.each(sourceDefinition._meteorMethods, function(method, methodName) {
+		targetDefinition._meteorMethods[methodName] = method;
 	});
 };
