+++ lib/modules/meteorMethods/hooks/parse_definition.js
@@ -36,7 +36,7 @@ function onParseDefinition(parsedDefinition, definition, className) {
                 ]);
             }
 
-            parsedDefinition.meteorMethods[meteorMethodName] = meteorMethod;
+            parsedDefinition._meteorMethods[meteorMethodName] = meteorMethod;
         });
     }
 };
